Compress weights to uint8
