+++ tests/onnx/quantization/test_calculation_quantizer_params.py
@@ -12,9 +12,9 @@
 import numpy as np
 import pytest
 
+from nncf.onnx.quantization.quantizer_parameters import calculate_scale_zero_point
 from nncf.onnx.quantization.quantizer_parameters import get_level_low_level_high
 from nncf.onnx.statistics.collectors import ONNXMinMaxTensorStatistic
-from nncf.quantization.fake_quantize import calculate_scale_zero_point
 from tests.post_training.test_templates.test_calculate_quantizer_parameters import TemplateTestFQParams
